BreastCancer/src/ML/requirements.txt
flask==3.1.0
flask-cors==4.0.0
gunicorn==24.0.0
mysql-connector-python==8.2.0
joblib==1.2.0
scikit-learn==1.5.2
numpy==1.24.4
pandas==2.1.4
pyyaml==6.0.2
pyjwt==2.8.0
python-dotenv
torch
